+++ oshi-core/src/main/java/oshi/hardware/platform/windows/WindowsUsbDevice.java
@@ -25,34 +25,20 @@ package oshi.hardware.platform.windows;
 
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Map.Entry;
 import java.util.Set;
+import java.util.stream.Collectors;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import com.sun.jna.platform.win32.Cfgmgr32; // NOSONAR squid:S1191
-import com.sun.jna.platform.win32.Cfgmgr32Util;
-import com.sun.jna.platform.win32.COM.COMException;
-import com.sun.jna.platform.win32.COM.WbemcliUtil.WmiResult;
-import com.sun.jna.ptr.IntByReference;
+import com.sun.jna.platform.win32.Guid.GUID;
 
 import oshi.annotation.concurrent.Immutable;
-import oshi.driver.windows.wmi.Win32DiskDrive;
-import oshi.driver.windows.wmi.Win32DiskDrive.DeviceIdProperty;
-import oshi.driver.windows.wmi.Win32PnPEntity;
-import oshi.driver.windows.wmi.Win32PnPEntity.PnPEntityProperty;
-import oshi.driver.windows.wmi.Win32USBController;
-import oshi.driver.windows.wmi.Win32USBController.USBControllerProperty;
+import oshi.driver.windows.DeviceTree;
 import oshi.hardware.UsbDevice;
 import oshi.hardware.common.AbstractUsbDevice;
 import oshi.util.ParseUtil;
-import oshi.util.platform.windows.WmiQueryHandler;
-import oshi.util.platform.windows.WmiUtil;
-import oshi.util.tuples.Pair;
+import oshi.util.tuples.Quintet;
 import oshi.util.tuples.Triplet;
 
 /**
@@ -61,7 +47,8 @@ import oshi.util.tuples.Triplet;
 @Immutable
 public class WindowsUsbDevice extends AbstractUsbDevice {
 
-    private static final Logger LOG = LoggerFactory.getLogger(WindowsUsbDevice.class);
+    private static final GUID GUID_DEVINTERFACE_USB_HOST_CONTROLLER = new GUID(
+            "{3ABF6F2D-71C4-462A-8A92-1E6861E6AF27}");
 
     public WindowsUsbDevice(String name, String vendor, String vendorId, String productId, String serialNumber,
             String uniqueDeviceId, List<UsbDevice> connectedDevices) {
@@ -84,7 +71,7 @@ public class WindowsUsbDevice extends AbstractUsbDevice {
      * @return a list of {@link oshi.hardware.UsbDevice} objects.
      */
     public static List<UsbDevice> getUsbDevices(boolean tree) {
-        List<UsbDevice> devices = getUsbDevices();
+        List<UsbDevice> devices = queryUsbDevices();
         if (tree) {
             return devices;
         }
@@ -92,47 +79,10 @@ public class WindowsUsbDevice extends AbstractUsbDevice {
         // Top level is controllers; they won't be added to the list, but all
         // their connected devices will be
         for (UsbDevice device : devices) {
+            // Recursively add all child devices
             addDevicesToList(deviceList, device.getConnectedDevices());
         }
         return deviceList;
-    }
-
-    private static List<UsbDevice> getUsbDevices() {
-        WmiQueryHandler h = WmiQueryHandler.createInstance();
-        boolean comInit = false;
-        try {
-            comInit = h.initCOM();
-
-            // Map to build the recursive tree structure
-            Map<String, List<String>> deviceTreeMap = new HashMap<>();
-            // Track devices seen in the process
-            Set<String> devicesSeen = new HashSet<>();
-
-            // Navigate the device tree to track what devices are present
-            List<UsbDevice> controllerDevices = new ArrayList<>();
-            List<String> controllerDeviceIdList = getControllerDeviceIdList(h);
-            for (String controllerDeviceId : controllerDeviceIdList) {
-                putChildrenInDeviceTree(controllerDeviceId, 0, deviceTreeMap, devicesSeen);
-            }
-            // Map to store information using PNPDeviceID as the key.
-            Map<String, Triplet<String, String, String>> deviceStringMap = queryDeviceStringsMap(h, devicesSeen);
-            // recursively build results
-            for (String controllerDeviceId : controllerDeviceIdList) {
-                WindowsUsbDevice deviceAndChildren = getDeviceAndChildren(controllerDeviceId, "0000", "0000",
-                        deviceTreeMap, deviceStringMap);
-                if (deviceAndChildren != null) {
-                    controllerDevices.add(deviceAndChildren);
-                }
-            }
-            return controllerDevices;
-        } catch (COMException e) {
-            LOG.warn("COM exception: {}", e.getMessage());
-            return Collections.emptyList();
-        } finally {
-            if (comInit) {
-                h.unInitCOM();
-            }
-        }
     }
 
     private static void addDevicesToList(List<UsbDevice> deviceList, List<UsbDevice> list) {
@@ -144,141 +94,67 @@ public class WindowsUsbDevice extends AbstractUsbDevice {
         }
     }
 
-    private static Map<String, Triplet<String, String, String>> queryDeviceStringsMap(WmiQueryHandler h,
-            Set<String> devicesToAdd) {
-        if (devicesToAdd.isEmpty()) {
-            return Collections.emptyMap();
-        }
-        Map<String, Triplet<String, String, String>> deviceStringCache = new HashMap<>();
-        // Add devices not in the tree
-        StringBuilder sb = new StringBuilder();
-        boolean first = true;
-        for (String deviceID : devicesToAdd) {
-            if (first) {
-                sb.append(" WHERE (PnPDeviceID=\"");
-                first = false;
-            } else {
-                sb.append(" OR (PnPDeviceID=\"");
-            }
-            sb.append(deviceID).append("\")");
-        }
-        String whereClause = sb.toString();
-        // Get serial # for disk drives or other physical media
-        Map<String, String> pnpToSerialMap = new HashMap<>();
-        WmiResult<DeviceIdProperty> serialNumbers = Win32DiskDrive.queryDiskDriveId(h, whereClause);
-        for (int i = 0; i < serialNumbers.getResultCount(); i++) {
-            String pnpDeviceID = WmiUtil.getString(serialNumbers, DeviceIdProperty.PNPDEVICEID, i);
-            pnpToSerialMap.put(pnpDeviceID,
-                    ParseUtil.hexStringToString(WmiUtil.getString(serialNumbers, DeviceIdProperty.SERIALNUMBER, i)));
-        }
-        // Query Win32_PnPEntity to populate the maps
-        WmiResult<PnPEntityProperty> pnpEntity = Win32PnPEntity.queryDeviceId(h, whereClause);
-        for (int i = 0; i < pnpEntity.getResultCount(); i++) {
-            String pnpDeviceID = WmiUtil.getString(pnpEntity, PnPEntityProperty.PNPDEVICEID, i);
-            String name = WmiUtil.getString(pnpEntity, PnPEntityProperty.NAME, i);
-            String vendor = WmiUtil.getString(pnpEntity, PnPEntityProperty.MANUFACTURER, i);
-            deviceStringCache.put(pnpDeviceID,
-                    new Triplet<>(name, vendor, pnpToSerialMap.getOrDefault(pnpDeviceID, "")));
-            LOG.debug("Adding {} to USB device cache.", pnpDeviceID);
-        }
+    private static List<UsbDevice> queryUsbDevices() {
+        Quintet<Set<Integer>, Map<Integer, Integer>, Map<Integer, String>, Map<Integer, String>, Map<Integer, String>> controllerDevices = DeviceTree
+                .queryDeviceTree(GUID_DEVINTERFACE_USB_HOST_CONTROLLER);
+        Map<Integer, Integer> parentMap = controllerDevices.getB();
+        Map<Integer, String> nameMap = controllerDevices.getC();
+        Map<Integer, String> deviceIdMap = controllerDevices.getD();
+        Map<Integer, String> mfgMap = controllerDevices.getE();
 
-        return deviceStringCache;
-    }
-
-    /**
-     * Navigates the Device Tree to place all children PNPDeviceIDs into the map for
-     * the specified deviceID. Recursively adds children's children, etc.
-     *
-     * @param deviceId
-     *            The device to add respective children to the map
-     * @param deviceInstance
-     *            The device instance (devnode handle), if known. If set to 0, the
-     *            code will search for a match.
-     * @param deviceTreeMap
-     *            The overall device tree map that starts at the controllers
-     * @param devicesSeen
-     *            Devices we've seen so we can add/remove from cache later
-     */
-    private static void putChildrenInDeviceTree(String deviceId, int deviceInstance,
-            Map<String, List<String>> deviceTreeMap, Set<String> devicesSeen) {
-        devicesSeen.add(deviceId);
-        // If no devInst provided, find it
-        int devInst = deviceInstance;
-        if (devInst == 0) {
-            IntByReference pdnDevInst = new IntByReference();
-            Cfgmgr32.INSTANCE.CM_Locate_DevNode(pdnDevInst, deviceId, 0);
-            devInst = pdnDevInst.getValue();
-        }
-        // Now iterate the children. Call CM_Get_Child to get first child
-        IntByReference child = new IntByReference();
-        if (0 == Cfgmgr32.INSTANCE.CM_Get_Child(child, devInst, 0)) {
-            // Add first child to a list
-            List<String> childList = new ArrayList<>();
-            String childId = Cfgmgr32Util.CM_Get_Device_ID(child.getValue());
-            childList.add(childId);
-            deviceTreeMap.put(deviceId, childList);
-            putChildrenInDeviceTree(childId, child.getValue(), deviceTreeMap, devicesSeen);
-            // Find any other children
-            IntByReference sibling = new IntByReference();
-            while (0 == Cfgmgr32.INSTANCE.CM_Get_Sibling(sibling, child.getValue(), 0)) {
-                // Add to the list
-                String siblingId = Cfgmgr32Util.CM_Get_Device_ID(sibling.getValue());
-                deviceTreeMap.get(deviceId).add(siblingId);
-                putChildrenInDeviceTree(siblingId, sibling.getValue(), deviceTreeMap, devicesSeen);
-                // Make this sibling the new child to find other siblings
-                child = sibling;
-            }
-        }
-    }
-
-    private static WindowsUsbDevice getDeviceAndChildren(String hubDeviceId, String vid, String pid,
-            Map<String, List<String>> deviceTreeMap, Map<String, Triplet<String, String, String>> deviceStringMap) {
-        String vendorId = vid;
-        String productId = pid;
-        Pair<String, String> idPair = ParseUtil.parsePnPDeviceIdToVendorProductId(hubDeviceId);
-        if (idPair != null) {
-            vendorId = idPair.getA();
-            productId = idPair.getB();
-        }
-        List<String> pnpDeviceIds = deviceTreeMap.getOrDefault(hubDeviceId, new ArrayList<>());
         List<UsbDevice> usbDevices = new ArrayList<>();
-        for (String pnpDeviceId : pnpDeviceIds) {
-            WindowsUsbDevice deviceAndChildren = getDeviceAndChildren(pnpDeviceId, vendorId, productId, deviceTreeMap,
-                    deviceStringMap);
+        // recursively build results
+        for (Integer controllerDevice : controllerDevices.getA()) {
+            WindowsUsbDevice deviceAndChildren = queryDeviceAndChildren(controllerDevice, parentMap, nameMap,
+                    deviceIdMap, mfgMap, "0000", "0000", "");
             if (deviceAndChildren != null) {
                 usbDevices.add(deviceAndChildren);
             }
         }
-        Collections.sort(usbDevices);
-        if (deviceStringMap.containsKey(hubDeviceId)) {
-            // name, vendor, serial
-            Triplet<String, String, String> device = deviceStringMap.get(hubDeviceId);
-            String name = device.getA();
+        return usbDevices;
+    }
+
+    private static WindowsUsbDevice queryDeviceAndChildren(Integer device, Map<Integer, Integer> parentMap,
+            Map<Integer, String> nameMap, Map<Integer, String> deviceIdMap, Map<Integer, String> mfgMap, String vid,
+            String pid, String parentSerial) {
+        // Parse vendor and product IDs from the device ID
+        // If this doesn't work, use the IDs from the parent
+        String vendorId = vid;
+        String productId = pid;
+        String serial = parentSerial;
+        Triplet<String, String, String> idsAndSerial = ParseUtil
+                .parseDeviceIdToVendorProductSerial(deviceIdMap.get(device));
+        if (idsAndSerial != null) {
+            vendorId = idsAndSerial.getA();
+            productId = idsAndSerial.getB();
+            serial = idsAndSerial.getC();
+            if (serial.isEmpty() && vendorId.equals(vid) && productId.equals(pid)) {
+                serial = parentSerial;
+            }
+        }
+        // Iterate the parent map looking for children
+        Set<Integer> childDeviceSet = parentMap.entrySet().stream().filter(e -> e.getValue().equals(device))
+                .map(Entry::getKey).collect(Collectors.toSet());
+        // Recursively find those children and put in a list
+        List<UsbDevice> childDevices = new ArrayList<>();
+        for (Integer child : childDeviceSet) {
+            WindowsUsbDevice deviceAndChildren = queryDeviceAndChildren(child, parentMap, nameMap, deviceIdMap, mfgMap,
+                    vendorId, productId, serial);
+            if (deviceAndChildren != null) {
+                childDevices.add(deviceAndChildren);
+            }
+        }
+        Collections.sort(childDevices);
+        // Finally construct the object and return
+        if (nameMap.containsKey(device)) {
+            String name = nameMap.get(device);
             if (name.isEmpty()) {
                 name = vendorId + ":" + productId;
             }
-            return new WindowsUsbDevice(name, device.getB(), vendorId, productId, device.getC(), hubDeviceId,
-                    usbDevices);
+            String deviceId = deviceIdMap.get(device);
+            String mfg = mfgMap.get(device);
+            return new WindowsUsbDevice(name, mfg, vendorId, productId, serial, deviceId, childDevices);
         }
         return null;
-    }
-
-    /**
-     * Queries the USB Controller list
-     *
-     * @param h
-     *            An instantiated {@link WmiQueryHandler}. User should have already
-     *            initialized COM.
-     * @return A list of Strings of USB Controller PNPDeviceIDs
-     */
-    private static List<String> getControllerDeviceIdList(WmiQueryHandler h) {
-        List<String> controllerDeviceIdsList = new ArrayList<>();
-        // One time lookup of USB Controller PnP Device IDs which don't
-        // change
-        WmiResult<USBControllerProperty> usbController = Win32USBController.queryUSBControllers(h);
-        for (int i = 0; i < usbController.getResultCount(); i++) {
-            controllerDeviceIdsList.add(WmiUtil.getString(usbController, USBControllerProperty.PNPDEVICEID, i));
-        }
-        return controllerDeviceIdsList;
     }
 }